requirements.txt
Django
gunicorn
postgres
psycopg2
dj-database-url
django_heroku
django-crispy-forms
djangorestframework 